test: Create fixtures for STAC metadata
