+++ scripts/stac/imagery/tests/collection_test.py
@@ -15,10 +15,10 @@ from scripts.files.files_helper import ContentType
 from scripts.files.fs import read
 from scripts.files.fs_s3 import write
 from scripts.stac.imagery.collection import ImageryCollection
-from scripts.stac.imagery.item import ImageryItem, STACAsset
+from scripts.stac.imagery.item import ImageryItem, STACAsset, STACProcessing
 from scripts.stac.imagery.metadata_constants import CollectionMetadata
 from scripts.stac.imagery.provider import Provider, ProviderRole
-from scripts.stac.imagery.tests.generators import any_stac_processing, fixed_now_function
+from scripts.stac.imagery.tests.generators import fixed_now_function
 from scripts.stac.util.stac_extensions import StacExtensions
 from scripts.tests.datetimes_test import any_epoch_datetime, any_epoch_datetime_string
 
@@ -102,7 +102,9 @@ def test_interval_updated_from_existing(fake_collection_metadata: CollectionMeta
     assert collection.stac["extent"]["temporal"]["interval"] == [["2021-01-27T00:00:00Z", "2021-02-20T00:00:00Z"]]
 
 
-def test_add_item(fake_collection_metadata: CollectionMetadata, subtests: SubTests) -> None:
+def test_add_item(
+    any_stac_processing: STACProcessing, fake_collection_metadata: CollectionMetadata, subtests: SubTests
+) -> None:
     now = any_epoch_datetime()
     now_string = format_rfc_3339_datetime_string(now)
     collection = ImageryCollection(fake_collection_metadata, fixed_now_function(now))
@@ -119,7 +121,7 @@ def test_add_item(fake_collection_metadata: CollectionMetadata, subtests: SubTes
                 "updated": asset_updated_datetime,
             }
         ),
-        any_stac_processing(),
+        any_stac_processing,
     )
     geometry = {
         "type": "Polygon",
