+++ scripts/tests/collection_from_items_test.py
@@ -16,17 +16,16 @@ from scripts.datetimes import utc_now
 from scripts.files.fs_s3 import write
 from scripts.json_codec import dict_to_json_bytes
 from scripts.stac.imagery.collection import ImageryCollection
-from scripts.stac.imagery.item import ImageryItem
+from scripts.stac.imagery.item import ImageryItem, STACAsset, STACProcessing
 from scripts.stac.imagery.metadata_constants import CollectionMetadata
-from scripts.stac.imagery.tests.generators import any_stac_asset, any_stac_processing
 from scripts.tests.datetimes_test import any_epoch_datetime_string
 
 
 @pytest.fixture(name="item", autouse=True)
-def setup() -> Iterator[ImageryItem]:
+def setup(any_stac_asset: STACAsset, any_stac_processing: STACProcessing) -> Iterator[ImageryItem]:
     # Create mocked STAC Item
     with patch.dict(environ, {"GIT_HASH": "any Git hash", "GIT_VERSION": "any Git version"}):
-        item = ImageryItem("123", any_epoch_datetime_string(), any_stac_asset(), any_stac_processing())
+        item = ImageryItem("123", any_epoch_datetime_string(), any_stac_asset, any_stac_processing)
     geometry = {
         "type": "Polygon",
         "coordinates": [[1799667.5, 5815977.0], [1800422.5, 5815977.0], [1800422.5, 5814986.0], [1799667.5, 5814986.0]],
